requirements.txt
lightning
pandas
pillow
pycocotools
torch
torchvision

pyobjc-framework-Quartz

numpy<2